warning fixes
git-svn-id: http://svn.miranda-ng.org/main/trunk@13738 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/BuddyPounce/src/main.cpp
@@ -76,7 +76,7 @@ int MsgAck(WPARAM, LPARAM lParam)
 				dbei.flags = DBEF_UTF | DBEF_SENT;
 				dbei.szModule = (char*)ack->szModule;
 				dbei.timestamp = time(NULL);
-				dbei.cbBlob = mir_strlen(pszUtf) + 1;
+				dbei.cbBlob = (int)mir_strlen(pszUtf) + 1;
 				dbei.pBlob = (PBYTE)(char*)pszUtf;
 				db_event_add(ack->hContact, &dbei);
 			}
